Improvements to the History, AntiSpam, VendorEngrams, and Mute commands, along with general improvements.
**General:**

- Bot keeps track of role changes and if the Moderators role is added or removed to anyone it updats the database with the `is_mod` status to true/false depending. This will be used in conjunction with the Clan Admin site, as well as other future stuff and things.
- `.say` now requires a Channel ID before sending a mesage. Example `.say 363462844693479434 Testing 123`. This way you can send from a different channel like #mod-spam. you can still send from the same channel, but you need to use the Channel ID of that channel still.
- Fixed `FTuesday`.

**AntiSpam:**

- Discord Invite AntiSpam tool now whitelists `blog.discordapp.com` and won't falsely trigger.

**History:**

- Now shows last 25 history records instead of 5
- Also shows first 750 characters of note instead of ~150

**Vendor Engrams:**

- Changed from saying 300 PL to 330 PL

**Mute:**

- Updates the mute workflow as the events happen (Set mute role, log to database, etc) so you know where it's at in the process.

+++ src/commands/modtools/Say.ts
@@ -12,8 +12,8 @@ export default class Mute extends Command<SweeperClient> {
 	public constructor() {
 		super({
 			name: 'say',
-			desc: 'Send a message as the bot in same channel.',
-			usage: '<prefix>say <Message>?',
+			desc: 'Send a message as the bot the specified channel.',
+			usage: '<prefix>say <ChannelID> <Message>?',
 			group: 'modtools',
 			guildOnly: true,
 			callerPermissions: ['MANAGE_GUILD']
@@ -22,7 +22,12 @@ export default class Mute extends Command<SweeperClient> {
 
 	public async action(message: Message, args: Array<any>): Promise<any> {
 
+		let channel: TextChannel;
 		if (args[0]) {
+			channel = <TextChannel> await this.client.channels.get(args[0]);
+		}
+
+		if (args[1]) {
 			// Set note info
 			let messageData: string = '';
 			messageData = this.parseNote(args);
@@ -31,8 +36,10 @@ export default class Mute extends Command<SweeperClient> {
 			}
 
 			try {
-				message.channel.send(`${messageData}`);
-				message.delete();
+				if (channel) {
+					channel.send(`${messageData}`);
+					message.delete();
+				}
 			} catch (err) {
 				const modChannel: TextChannel = <TextChannel> message.guild.channels.get(Constants.modChannelId);
 				modChannel.send(`There was an error using the **say** command.\n\n**Error:**\n${err}`);
@@ -47,7 +54,7 @@ export default class Mute extends Command<SweeperClient> {
 
 	private parseNote(args: Array<any>): string {
 		let text: string = '';
-		for (let index: number = 0; index < args.length; index++) {
+		for (let index: number = 1; index < args.length; index++) {
 			text += `${args[index].trim()} `;
 		}
 		return text.slice(0, -1);
